pyproject.toml
[project]
name = "recipe"
version = "0.1.0"
description = ""
authors = [
    {name = "Eren Arıman",email = "arimaneren@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "orjson (>=3.11.7,<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "sentry-sdk (>=2.57.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "import-linter (>=2.9,<3.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
